This commit was manufactured by cvs2svn to create branch 'release121'.
git-svn-id: https://svn.java.net/svn/visualvm~svn/branches/release121@2314 ba0df5e5-9053-6d9f-941f-f112716aa91a

+++ visualvm/charts/src/com/sun/tools/visualvm/charts/SimpleXYChartSupport.java
@@ -187,7 +187,7 @@ public final class SimpleXYChartSupport {
      * @return formatted value in the same format as used in the chart
      */
     public String formatBytes(long value) {
-        return SimpleXYChartUtils.formatBytes((int)(value * chartFactor));
+        return SimpleXYChartUtils.formatBytes((long)(value * chartFactor));
     }
 
     /**
